Improve CreatePost & setPost. Create utilities folder and add date.ts
